requirements.txt
# Core Dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.24.3

# AI/ML Libraries
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.13
langchain-text-splitters==0.0.1
langchain-core==0.1.10

# Embeddings & Vector Store
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# Document Processing
pypdf==4.0.1
openpyxl==3.1.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0
